Remove --force-reinstall from pip cmd
We have bumpped version for vapi-client-bindings from 1.3.0 to 3.0.0. Note that there is no incompatible API changes in the new API bindings. The change is mainly because we had published 2.5 and 2.6 versions long time ago. In order to avoid confusion we decided to bump the version to 3.0.0 so a newer version can be upgraded even you have 2.5/2.6 installed locally. Hence the --force-reinstall arguments is no longer needed.